+++ riemann/tx/sapling.py
@@ -1,0 +1,9 @@
+# import riemann
+# from riemann import utils
+# from riemann.tx import ByteData, VarInt
+# from riemann.script import serialization
+
+SIGHASH_ALL = 0x01
+SIGHASH_NONE = 0x02
+SIGHASH_SINGLE = 0x03
+SIGHASH_ANYONECANPAY = 0x80
